Switch to Memcache PHP-extenstion instead of Memcached
There seems to be some issues with object splitting when using the
Memcached extension and the Drupal integration module. This results
and cache misses and degraded performance.

For more info about the issues:
https://www.drupal.org/project/memcache/issues/2864659
https://www.drupal.org/project/memcache/issues/2677660

Also there might be a solution to the Memcached extension issue,
by enabling compression by default:

https://www.drupal.org/project/memcache/issues/2958403

But we're lucky that there was just recently released a PHP 7
compatible version of Memcache extension so we just use that :)